requirements.txt
mss>=9.0.0
opencv-python>=4.8.0
pyaudio>=0.2.14
numpy>=1.24.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
psutil>=5.9.0

